Add "controller" functionality. (#69)
We tap into the existing control channel and provide a top-level method
on the receiver to get a controller instance, similar to a sender or
clocksource instance.

The controller can then send commands out-of-band to the receiver.
Currently, the only supported command is the "snapshot meters" command,
which lets you get a ... snapshot of the meters as they exist! :)

This allows the receiver to not have to know or care about what
reporting endpoints are tapped into it, or care itself about running the
HTTP endpoint.  It can utilize the same high-speed event polling
mechanism without requiring references to the receiver directly.

The HTTP server has be renamed/moved into tic::http::HttpReporter, and
runs standalone from the receiver, but can be plumbed together via
Receiver::get_controller().  Both example projects show how this now
looks.

Overall, throughput increased slightly -- 1% to 2% on average -- and
latency p50 dropped about 6% to 7% on average.  This is in the
uncontended case when nobody is actually querying the HTTP endpoint for
the metrics.  Obviously, if you do show, the receiver has to do
non-metrics-ingestion work, but it's incredibly minimal.